created folder AdminModals and moved AdminEditModal and AdminSuspendModal in it
