+++ frontend/src/components/AdminControls/Users.js
@@ -4,8 +4,8 @@ import BootstrapTable from 'react-bootstrap-table-next';
 import paginationFactory from 'react-bootstrap-table2-paginator';
 import { Type } from 'react-bootstrap-table2-editor';
 import './AdminTables.css';
-import AdminEditModal from '../AdminEditModal/AdminEditModal';
-import AdminSuspendModal from '../AdminSuspendModal/AdminSuspendModal';
+import AdminEditModal from '../AdminModals/AdminEditModal';
+import AdminSuspendModal from '../AdminModals/AdminSuspendModal';
 
 const Users = ({ users, refreshUsers }) => {
   const [editModal, setEditModal] = useState(false);
